refactor: move distance functions to sequence.py
